requirements.txt
openai-whisper
typer
flake8
torch
moviepy
ffmpeg-python
torchaudio
soundfile
